Use unsigned long long to store unix timestamp millis
Before long was used which caused an overflow on Raspberry Pi 4.

long is ambiguously defined as 'at least 4 bytes'. Implementation is system-dependent.
unsigned long long is defined to always have 8 bytes.

+++ sender-receiver-linux/Logging/LogEntries/ActuatorLogEntry.cpp
@@ -61,7 +61,7 @@ ActuatorLogEntry::ActuatorLogEntry(const time_point<system_clock> &timePoint, un
 
 // JSON converter class (used implicitly by nlohmann/json.hpp):
 void to_json(nlohmann::json &jsonObject, const ActuatorLogEntry &entry) {
-    long timePointAsUnixMillis = std::chrono::duration_cast<std::chrono::milliseconds>(
+    unsigned long long timePointAsUnixMillis = std::chrono::duration_cast<std::chrono::milliseconds>(
             entry.getTimePoint().time_since_epoch()).count();
 
     jsonObject = {
